requirements-test.txt
# requirements of main dependencies used for tests
pyramid==1.10.5
pyramid-tm==2.4
transaction==3.0.1
requests==2.25.1
requests-mock==1.8.0
sqlalchemy==1.3.22
python-slugify==4.0.1
text-unidecode==1.3
zope.interface==5.2.0
zope.sqlalchemy==1.3
inflect==5.0.2
# tests
pytest==6.2.2
pytest-cov==2.11.1
coverage==5.4
mock==4.0.3; python_version>'3'
typing-extensions==3.7.4.3; python_version<'3.8'
-e .[tests]
